SwordAttack
I have Implemented a sword attack that is using line traces that are being called only at a certain frames of the attack animation. This can work with different attacks but you just need to update the AttackAnimationNotifiers and the end Attack Notifier on the different attack animations. Each weapon needs to have 2 sockets which are used for the line traces, 1 called Start and 1 End, between these 2 sockets the player casts the line trace that checks for hits.